change delay to take a float instead of an int, makes it more precise

+++ source/libarduino/inc/io/io.hpp
@@ -56,6 +56,6 @@ void analogReference(VOLTTYPE);
 void analogWrite(PIN, PWMVALUE);
 
 // MISC
-void delay(int);
+void delay(float);
 
 #endif
